code review comments addressed #4
