+++ tools/content-loader/src/helpers/memoized-read.ts
@@ -1,0 +1,12 @@
+export type Read = (path: string) => Promise<any>;
+
+export const memoizeRead = (f: Read) => {
+  const cache: any = {};
+  return async (path: string) => {
+    const cachedResult = cache[path];
+    if (!cachedResult) {
+      cache[path] = f(path);
+    }
+    return cache[path];
+  };
+};
