+++ tools/content-loader/src/helpers/read-yaml.ts
@@ -13,16 +13,3 @@ export const readYaml = async (file: string): Promise<any> => {
 
   return safeLoad(data);
 };
-
-export type Read = (path: string) => Promise<any>;
-
-export const memoizeRead = (f: Read) => {
-  const cache: any = {};
-  return async (path: string) => {
-    const cachedResult = cache[path];
-    if (!cachedResult) {
-      cache[path] = f(path);
-    }
-    return cache[path];
-  };
-};
